test(pacquet): move directory fetcher tests out of line
